Middleware to Niraffe library
