+++ niraffe/index.ts
@@ -1,5 +1,5 @@
 export * from "./core";
 export * from "./httpStatusCodeHandlers";
 export * from "./negotiation";
-//export * from "./endpointRouting";
 export * from "./routing";
+export * from "./middleware";
